Bug 1822446 - Add OnboardingController for handling onboarding functionality
