+++ fenix/app/src/test/java/org/mozilla/fenix/home/recentvisits/interactor/RecentVisitsInteractorTest.kt
@@ -20,6 +20,7 @@ import org.mozilla.fenix.home.recentvisits.RecentlyVisitedItem.RecentHistoryHigh
 import org.mozilla.fenix.home.recentvisits.controller.RecentVisitsController
 import org.mozilla.fenix.home.sessioncontrol.DefaultSessionControlController
 import org.mozilla.fenix.home.sessioncontrol.SessionControlInteractor
+import org.mozilla.fenix.onboarding.controller.OnboardingController
 
 class RecentVisitsInteractorTest {
     private val defaultSessionControlController: DefaultSessionControlController =
@@ -29,6 +30,7 @@ class RecentVisitsInteractorTest {
     private val recentBookmarksController: RecentBookmarksController = mockk(relaxed = true)
     private val pocketStoriesController: PocketStoriesController = mockk(relaxed = true)
     private val recentVisitsController: RecentVisitsController = mockk(relaxed = true)
+    private val onboardingController: OnboardingController = mockk(relaxed = true)
 
     private lateinit var interactor: SessionControlInteractor
 
@@ -41,6 +43,7 @@ class RecentVisitsInteractorTest {
             recentBookmarksController,
             recentVisitsController,
             pocketStoriesController,
+            onboardingController,
         )
     }
